[python]: clean up a bit the code
